Extract valid_chain_start?/3 pipe helper function into helper module
